+++ src/blocks/button-maxi/save.js
@@ -7,8 +7,9 @@ import { RawHTML } from '@wordpress/element';
  * Internal dependencies
  */
 import { Button } from '../../components';
-import MaxiBlock, { getMaxiBlockAttributes } from '../../components/maxi-block';
+import MaxiBlock from '../../components/maxi-block';
 
+import { getMaxiBlockAttributes } from '../../extensions/maxi-block';
 /**
  * External dependencies
  */
