+++ src/blocks/image-maxi/edit.js
@@ -16,8 +16,12 @@ import {
 	getGroupAttributes,
 	getLastBreakpointAttribute,
 } from '../../extensions/styles';
-import MaxiBlock, { getMaxiBlockAttributes } from '../../components/maxi-block';
-import { MaxiBlockComponent } from '../../extensions/maxi-block';
+import MaxiBlock from '../../components/maxi-block';
+import {
+	getMaxiBlockAttributes,
+	MaxiBlockComponent,
+} from '../../extensions/maxi-block';
+
 import {
 	BlockResizer,
 	Button,
