+++ src/blocks/number-counter-maxi/save.js
@@ -6,7 +6,8 @@ eslint-disable react/no-unknown-property
 /**
  * Internal dependencies
  */
-import MaxiBlock, { getMaxiBlockAttributes } from '../../components/maxi-block';
+import MaxiBlock from '../../components/maxi-block';
+import { getMaxiBlockAttributes } from '../../extensions/maxi-block';
 
 /**
  * External dependencies
