+++ src/blocks/svg-icon-maxi/edit.js
@@ -12,10 +12,11 @@ import Inspector from './inspector';
 import {
 	getResizerSize,
 	MaxiBlockComponent,
+	getMaxiBlockAttributes,
 } from '../../extensions/maxi-block';
 import { Toolbar, BlockResizer, RawHTML } from '../../components';
 import { getLastBreakpointAttribute } from '../../extensions/styles';
-import MaxiBlock, { getMaxiBlockAttributes } from '../../components/maxi-block';
+import MaxiBlock from '../../components/maxi-block';
 import MaxiModal from '../../editor/library/modal';
 import getStyles from './styles';
