+++ src/extensions/maxi-block/index.js
@@ -1,2 +1,3 @@
 export { default as MaxiBlockComponent } from './maxiBlockComponent';
+export { default as getMaxiBlockAttributes } from './getMaxiBlockAttributes';
 export * from './utils';
